+++ apps/backoffice/CHANGELOG.md
@@ -1,5 +1,17 @@
 # io-services-cms-backoffice
 
+## 1.5.1
+
+### Patch Changes
+
+- ea9b420: fix topic id null check
+
+## 1.5.0
+
+### Minor Changes
+
+- 42c38d0: Add frontend topic management on service create/update
+
 ## 1.4.6
 
 ### Patch Changes