+++ apps/io-services-cms-webapp/CHANGELOG.md
@@ -1,5 +1,17 @@
 # io-services-cms-webapp
 
+## 1.25.2
+
+### Patch Changes
+
+- ea9b420: fix topic id null check
+
+## 1.25.1
+
+### Patch Changes
+
+- c2ac0ea: fix query get topic
+
 ## 1.25.0
 
 ### Minor Changes